refactor(Refactored-routes-to-use-route-groups): Refactored routes to use route groups
