+++ app/(dashboard)/inventory/page.tsx
@@ -1,5 +1,5 @@
-
 'use client'
+
 import React from 'react'
 
 const page = () => {
