+++ app/(dashboard)/inventory/styles/inventory.css
